+++ src/Moq/Matchers/AnyMatcher.cs
@@ -12,5 +12,9 @@ namespace Moq.Matchers
 		}
 
 		public bool Matches(object value) => true;
-	}
+
+		public void SetupEvaluatedSuccessfully(object value)
+		{
+		}
+    }
 }