+++ src/Moq/Setup.cs
@@ -66,6 +66,12 @@ namespace Moq
 			}
 		}
 
+		public void EvaluatedSuccessfully(Invocation invocation)
+		{
+			this.Condition?.SetupEvaluatedSuccessfully();
+			this.expectation.SetupEvaluatedSuccessfully(invocation);
+		}
+
 		public virtual void SetOutParameters(Invocation invocation)
 		{
 		}